pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fluxflow"
version = "0.1.0"
description = "Core model and inference for FluxFlow text-to-image generation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "FluxFlow Contributors"}
]
keywords = ["deep-learning", "text-to-image", "diffusion", "vae", "transformers", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Generation",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "safetensors>=0.3.0",
    "transformers>=4.30.0",
    "diffusers>=0.20.0",
    "einops>=0.6.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0,<2.0",
    "matplotlib>=3.5.0",
    "orjson>=3.8.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.1.0",
    "pytest-mock>=3.11.0",
    # Code quality
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "mypy>=1.4.0",
    "ruff>=0.1.0",
    # Type stubs
    "types-pyyaml>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/danny-mio/fluxflow-core"
Repository = "https://github.com/danny-mio/fluxflow-core"
Documentation = "https://github.com/danny-mio/fluxflow-core/blob/main/README.md"
Issues = "https://github.com/danny-mio/fluxflow-core/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["fluxflow*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"fluxflow" = ["py.typed"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py310", "py311"]
include = '\.pyi?$'

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "transformers.*",
    "diffusers.*",
    "einops.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fluxflow.models.diffusion_pipeline"
disable_error_code = ["attr-defined", "arg-type", "assignment", "no-any-return"]

[[tool.mypy.overrides]]
module = "fluxflow.utils.io"
disable_error_code = ["assignment"]
